refactor to move d3 drawing and utility functions outside of react component
